# vlm: tensor hash kernel