--- a/python/sglang/srt/layers/multimodal.py
+++ b/python/sglang/srt/layers/multimodal.py
@@ -19,44 +19,52 @@ import triton.language as tl
 
 
 @triton.jit
-def simple_hash_kernel(
-    input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.constexpr, PRIME: tl.constexpr
+def hash_kernel(
+    input_ptr,
+    output_ptr,
+    n_elements,
+    BLOCK_SIZE: tl.constexpr,
+    PRIME: tl.constexpr,
+    XCONST: tl.constexpr,
 ):
     pid = tl.program_id(axis=0)
     block_start = pid * BLOCK_SIZE
     offsets = block_start + tl.arange(0, BLOCK_SIZE)
     mask = offsets < n_elements
 
-    # Load data
     data = tl.load(input_ptr + offsets, mask=mask, other=0)
+    mixed = data ^ (offsets + XCONST)
+    hash_val = mixed * PRIME
+    hash_val = hash_val ^ (hash_val >> 16)
+    hash_val = hash_val * (PRIME ^ XCONST)
+    hash_val = hash_val ^ (hash_val >> 13)
 
-    # Simple hash computation (multiplication and shifts, no xor/shuffle)
-    hash_val = data * PRIME
-    hash_val = hash_val + (hash_val >> 16)
-    hash_val = hash_val * PRIME
-    hash_val = hash_val + (hash_val >> 13)
-
-    # Store intermediate hash values
     tl.store(output_ptr + offsets, hash_val, mask=mask)
 
 
-def gpu_tensor_hash(tensor: torch.Tensor, prime=0x01000193) -> int:
-    assert tensor.is_cuda, "Tensor must be on CUDA device"
+PRIME_1 = -(11400714785074694791 ^ 0xFFFFFFFFFFFFFFFF) - 1
+PRIME_2 = -(14029467366897019727 ^ 0xFFFFFFFFFFFFFFFF) - 1
+
+
+def gpu_tensor_hash(tensor: torch.Tensor) -> int:
+    assert tensor.is_cuda
     tensor = tensor.contiguous().view(torch.int32)
-
-    n_elements = tensor.numel()
+    n = tensor.numel()
     BLOCK_SIZE = 1024
-    grid = (triton.cdiv(n_elements, BLOCK_SIZE),)
+    grid = (triton.cdiv(n, BLOCK_SIZE),)
 
-    intermediate_hashes = torch.empty(
-        n_elements, dtype=torch.int32, device=tensor.device
+    intermediate_hashes = torch.empty(n, dtype=torch.int32, device=tensor.device)
+
+    hash_kernel[grid](
+        tensor,
+        intermediate_hashes,
+        n,
+        BLOCK_SIZE=BLOCK_SIZE,
+        PRIME=PRIME_1,
+        XCONST=PRIME_2,
     )
 
-    simple_hash_kernel[grid](
-        tensor, intermediate_hashes, n_elements, BLOCK_SIZE=BLOCK_SIZE, PRIME=prime
-    )
-
-    # Final reduction on GPU (simple sum reduction)
+    # TODO: threads cant be synced on triton kernel
     final_hash = intermediate_hashes.sum().item()
 
     return final_hash