<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Previously, for gpu-tensor, the hash process required by multimodal model will first:
1. move the tensor D2H
2. then hash the tensor with sha256

With some simple profiling, hashing a normal image feature (e.g., from sgl-logo image, shape=[3312,1176], dtype=float32 in qwen-vl cases) would cost ~80ms.

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

1. Add a triton hash kernel, which does:
   1. hash the blocks parallelly on GPU
   2. reduce the block results sequentially on CPU

## Profiling

### Hash performance

|               | time(1000 times) |
|---------------|------------------|
| original hash | 11s              |
| triton hash   | ~0.6s            |

### MMMU
|                        | accuracy(before) | time(before) | accuracy(after) | time(after) |  
|------------------------|------------------|--------------|-----------------|-------------|
| Gemma-3-4b-it          | 0.384            | 226.2        | 0.384           | 218.3       | 
| Qwen2.5-VL-7B-Instruct | 0.467            | 352.7        | 0.467           | 338.2       | 
| Minicpmv               | 0.436            | 232.5        | 0.436           | 230.7       | 


## Correctness

1. **Consistency**: The kernel first hash tensor with blocks, then performs a reduce on block results sequentially.
4. **Collision**: The collision probability is zero, on 10000 randomly-generated tensors with the same shape as real data, with the following script:
```python

def test_hash_collision(hasher, name, num_tensors=10000, tensor_shape=(128,)):
    hashes = set()
    collision_count = 0

    for i in range(num_tensors):
        tensor = torch.rand(
         size=tensor_shape, dtype=torch.float32, device="cuda"
        ) * 2 - 1
        h = hasher(tensor)
        if h in hashes:
            collision_count += 1
        else:
            hashes.add(h)

    print(f"hasher: {name}")
    print(f"Tested {num_tensors} random tensors of shape {tensor_shape}")
    print(f"Collisions found: {collision_count}")
    print(f"Collision rate: {collision_count / num_tensors:.6f}")

```
<!-- Describe the changes made in this PR. -->

## Future Work

1.  Should we just implement MurmurHash/xxHash directly?

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.